# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API хакатон-проекта.",
        "title": "Hackaton API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/v1",
    "paths": {
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Проверка живости (API + БД)",
                "operationId": "health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HealthResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/httperr.Response"
                        }
                    }
                }
            }
        },
        "/items": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Список элементов",
                "operationId": "list-items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ListResponse-dto_Item"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Создать элемент",
                "operationId": "create-item",
                "parameters": [
                    {
                        "description": "Данные элемента",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.Item"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httperr.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/httperr.Response"
                        }
                    }
                }
            }
        },
        "/items/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Элемент по id",
                "operationId": "get-item",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID элемента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Item"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httperr.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httperr.Response"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "items"
                ],
                "summary": "Удалить элемент",
                "operationId": "delete-item",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID элемента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httperr.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httperr.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateItemRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1,
                    "example": "Первый элемент"
                }
            }
        },
        "dto.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "dto.Item": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.ListResponse-dto_Item": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Item"
                    }
                }
            }
        },
        "httperr.Error": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "поле запроса, если ошибка про него",
                    "type": "string"
                },
                "msg": {
                    "description": "человекочитаемое сообщение",
                    "type": "string",
                    "example": "Не найдено"
                },
                "status": {
                    "description": "машинный код для фронта",
                    "type": "string",
                    "example": "not_found"
                }
            }
        },
        "httperr.Response": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httperr.Error"
                    }
                }
            }
        }
    }
}